// src/models/view-related/image.model.ts
export interface DFDImages {
  interactorImage?: HTMLImageElement;
  processImage?: HTMLImageElement;
  dataStoreImage?: HTMLImageElement;
}

export interface DeleteImages {
  deleteImage?: HTMLImageElement;
  deleteHoverImage?: HTMLImageElement;
}

export interface Images {
  elementImages?: DFDImages;
  deleteImages?: DeleteImages;
  dataFlowTip?: HTMLImageElement;
}